create useSongProgress custom hook
